+++ packages/flame/lib/src/components/text_box_component.dart
@@ -189,7 +189,12 @@ class TextBoxComponent<T extends TextRenderer> extends PositionComponent {
     final recorder = PictureRecorder();
     final c = Canvas(recorder, size.toRect());
     _fullRender(c);
-    return recorder.endRecording().toImage(width.toInt(), height.toInt());
+    //final devicePixelRatio = window.devicePixelRatio;
+    final devicePixelRatio = window.devicePixelRatio;
+    return recorder.endRecording().toImage(
+          (width * devicePixelRatio).toInt(),
+          (height * devicePixelRatio).toInt(),
+        );
   }
 
   /// Override this method to provide a custom background to the text box.